### 2025-05-27 #8289 @editorialbot
[PRE REVIEW]: xr_fresh: Automated Time Series Feature Extraction for Remote Sensing & Gridded Data
**Submitting author:** @mmann1123 (<a href="http://orcid.org/0000-0002-6268-6867">Michael Mann</a>)
**Repository:** https://github.com/mmann1123/xr_fresh
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.2
