### 2025-05-27 #8289 @crvernon
:wave: @mmann1123 - Thanks for your submission to JOSS.  We are currently working through a large backlog of submission and will get this one assigned to a topic editor as soon as possible.

While you are waiting, please reduce the word count of your paper to <= 1000 words and handle any issues with DOI listed above.
