+++ writeup/paper.md
@@ -28,38 +28,12 @@ abstract: |
  
 <!-- compile working with:
 pandoc writeup.md --bibliography=refs.bib --filter pandoc-citeproc --pdf-engine=xelatex -o output.pdf-->
-
-<!-- pandoc writeup.md --template=mytemplate.tex -o output.pdf --bibliography=refs.bib --pdf-engine=xelatex --filter pandoc-citeproc -->
-
  
-<!-- convert to word doc (2 steps)  --template=mytemplate.tex \
-pandoc writeup.md 
-  --from markdown+raw_tex \
-  --to latex \
-  --bibliography=refs.bib \
-  --filter pandoc-citeproc\
-  -o output.tex
-
-pandoc output.tex --from latex --to docx -o output.docx
---> 
-<!-- 
-Look at https://mpastell.com/pweave/docs.html -->
-
 ## Statement of need
 
-Recent advancements in satellite technology and sensor data have significantly enhanced our capability to monitor environmental phenomena through time series analysis of images. Efficiently extracting relevant time series features at scale remains challenging, necessitating automated and robust methods. Inspired by `tsfresh` (Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests), we introduce `xr_fresh`, tailored specifically for image time series,by automating the extraction of time series features on a pixel-by-pixel basis [@CHRIST201872].
+Gridded time series data from satellites, climate models, camera feeds, and sensors contain rich temporal information for applications like crop type classification and yields, anomaly detection, robotics, quality control, environmental monitoring, and natural resource management [@delince2017handbook;@mumuni2024automated;@hufkens2019monitoring;@MANN201760;@mann2019predicting]. Efficiently extracting relevant time series features at scale remains challenging, necessitating automated [@faouzi2022time;@li2020forecasting]. Inspired by `tsfresh`, we introduce `xr_fresh`, tailored specifically for image time series,by automating the extraction of time series features on a pixel-by-pixel basis [@CHRIST201872].
 
-Time series data, characterized by sequences of measurements indexed over time, are ubiquitous in diverse fields, including finance, healthcare, industrial monitoring, and environmental sciences [@faouzi2022time]. Analyzing and modeling these temporal sequences for tasks such as classification (assigning a time series to a predefined category) or regression (predicting a continuous value associated with a time series) is a fundamental problem in machine learning [@faouzi2022time]. However, applying standard machine learning algorithms directly to raw time series data is often challenging due to the inherent temporal structure and potentially high dimensionality of the sequences [@faouzi2022time]. An effective representation of the time series characteristics is crucial for building accurate and robust models [@mumuni2024automated;@faouzi2022time].
-
-Traditional approaches to time series analysis and machine learning often rely on manual feature engineering, a process in which domain experts handcraft relevant features from raw data [@li2020forecasting]. Although effective in specific contexts, this process is labor intensive, requires significant expertise, and may not generalize well across different datasets or tasks [@li2020forecasting;@faouzi2022time]. The increasing volume and complexity of modern time series data require more automated and scalable approaches to data processing and feature extraction  [@mumuni2024automated;@li2020forecasting].
-
[…]
