+++ writeup/old_intro.md
@@ -1,0 +1,342 @@
+
+Recent advancements in satellite technology and sensor data have significantly enhanced our capability to monitor environmental phenomena through time series analysis of images. Efficiently extracting relevant time series features at scale remains challenging, necessitating automated and robust methods. Inspired by `tsfresh` (Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests), we introduce `xr_fresh`, tailored specifically for image time series,by automating the extraction of time series features on a pixel-by-pixel basis [@CHRIST201872].
+
+Time series data, characterized by sequences of measurements indexed over time, are ubiquitous in diverse fields, including finance, healthcare, industrial monitoring, and environmental sciences [@faouzi2022time]. Analyzing and modeling these temporal sequences for tasks such as classification (assigning a time series to a predefined category) or regression (predicting a continuous value associated with a time series) is a fundamental problem in machine learning [@faouzi2022time]. However, applying standard machine learning algorithms directly to raw time series data is often challenging due to the inherent temporal structure and potentially high dimensionality of the sequences [@faouzi2022time]. An effective representation of the time series characteristics is crucial for building accurate and robust models [@mumuni2024automated;@faouzi2022time].
+
+Traditional approaches to time series analysis and machine learning often rely on manual feature engineering, a process in which domain experts handcraft relevant features from raw data [@li2020forecasting]. Although effective in specific contexts, this process is labor intensive, requires significant expertise, and may not generalize well across different datasets or tasks [@li2020forecasting;@faouzi2022time]. The increasing volume and complexity of modern time series data require more automated and scalable approaches to data processing and feature extraction  [@mumuni2024automated;@li2020forecasting].
+
+A prominent tool in this domain is `tsfresh` [@CHRIST201872], a Python package specifically designed for time series feature extraction based on scalable hypothesis tests   [jin2022automated;@mumuni2024automated;@gilpin2021chaos;@schroeder2019chemiresistive;@sylligardos2023choose;@mcdermott2023event;@petropoulos2022forecasting;@zhao2022ai]. `tsfresh` automates the process of calculating a diverse and large number of features from time series data. These features are derived from existing research and include distribution properties (e.g., mean, standard deviation, kurtosis), measures of autocorrelation (e.g., fast Fourier transform (FFT) and power spectral density coefficients), properties of linear regressors (e.g., gradient, standard error), energy, entropy, and stochasticity [@CHRIST201872]. After feature extraction, `tsfresh` also offers automated feature selection methods, such as using statistical tests like the Mann-Whitney U test, to identify and select the most relevant features for a specific task [@CHRIST201872].
+
+Gridded time series data from satellites, climate models, camera feeds, and sensors contain rich temporal information for applications like crop type classification and yields, anomaly detection, robotics, quality control, environmental monitoring, and natural resource management [@delince2017handbook;@mumuni2024automated;@hufkens2019monitoring;@MANN201760;@mann2019predicting]. Traditional methods relied heavily on manually engineered features, which are time-consuming, require domain expertise, and often lack scalability.
+
[…]
