+++ writeup/refs.bib
@@ -1,14 +1,4 @@
-
-@article{begue2018remote,
-  title={Remote sensing and cropping practices: A review},
-  author={B{\'e}gu{\'e}, Agn{\`e}s and Arvor, Damien and Bellon, Beatriz and Betbeder, Julie and De Abelleyra, Diego and PD Ferraz, Rodrigo and Lebourgeois, Valentine and Lelong, Camille and Sim{\~o}es, Margareth and R. Ver{\'o}n, Santiago},
-  journal={Remote Sensing},
-  volume={10},
-  number={1},
-  pages={99},
-  year={2018},
-  publisher={MDPI}
-}
+ 
 
 @article{CHRIST201872,
 title = {Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh – A Python package)},
@@ -28,6 +18,7 @@ abstract = {Time series feature engineering is a time-consuming process because 
   title={Handbook on remote sensing for agricultural statistics},
   author={Delince, J and Lemoine, G and Defourny, P and Gallego, J and Davidson, A and Ray, S and Rojas, O and Latham, J and Achard, F},
   journal={GSARS: Rome, Italy},
+  doi={10.13140/RG.2.2.13259.69920},
   year={2017}
 }
 @article{faouzi2022time,
@@ -35,7 +26,9 @@ abstract = {Time series feature engineering is a time-consuming process because 
   author={Faouzi, Johann},
   journal={Machine Learning (Emerging Trends and Applications)},
   year={2022},
-  publisher={Proud Pen}
+  note = {hal-03558165},
+  doi = {10.5772/intechopen.1004810}
+
 }
 @article{funk2015climate,
   title={The climate hazards infrared precipitation with stations—a new environmental record for monitoring extremes},
@@ -45,23 +38,10 @@ abstract = {Time series feature engineering is a time-consuming process because 
   number={1},
   pages={1--21},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/sdata.2015.66}
 }
 
-
-@article{gilpin2021chaos,
-  title={Chaos as an interpretable benchmark for forecasting and data-driven modelling},
-  author={Gilpin, William},
-  journal={arXiv preprint arXiv:2110.05266},
-  year={2021}
-}
-@inproceedings{hohl2024recent,
-  title={Recent Trends Challenges and Limitations of Explainable AI in Remote Sensing},
-  author={H{\"o}hl, Adrian and Obadic, Ivica and Fern{\'a}ndez-Torres, Miguel-{\'A}ngel and Oliveira, Dario and Zhu, Xiao Xiang},
-  booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
-  pages={8199--8205},
-  year={2024}
-}
 @article{hufkens2019monitoring,
   title={Monitoring crop phenology using a smartphone based near-surface remote sensing approach},
   author={Hufkens, Koen and Melaas, Eli K and Mann, Michael L and Foster, Timothy and Ceballos, Francisco and Robles, Miguel and Kramer, Berber},
@@ -69,14 +49,15 @@ abstract = {Time series feature engineering is a time-consuming process because 
   volume={265},
   pages={327--337},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.agrformet.2018.11.002}
 }
-@misc{jax2018github,
+@software{jax2018github,
[…]
