### 2025-05-27 e7ade05
amend abstract
