+++ writeup/paper.md
@@ -31,7 +31,7 @@ pandoc writeup.md --bibliography=refs.bib --filter pandoc-citeproc --pdf-engine=
  
 ## Statement of need
 
-Gridded time series data from satellites, climate models, camera feeds, and sensors contain rich temporal information for applications like crop type classification and yields, anomaly detection, robotics, quality control, environmental monitoring, and natural resource management [@delince2017handbook;@mumuni2024automated;@hufkens2019monitoring;@MANN201760;@mann2019predicting]. Efficiently extracting relevant time series features at scale remains challenging, necessitating automated [@faouzi2022time;@li2020forecasting]. Inspired by `tsfresh`, we introduce `xr_fresh`, tailored specifically for image time series,by automating the extraction of time series features on a pixel-by-pixel basis [@CHRIST201872].
+Gridded time series data from satellites, climate models, camera feeds, and sensors contain rich temporal information for applications like crop type classification and yields, anomaly detection, robotics, quality control, environmental monitoring, and natural resource management [@delince2017handbook;@mumuni2024automated;@hufkens2019monitoring;@MANN201760;@mann2019predicting]. Efficiently extracting relevant time series features at scale remains challenging, necessitating automation [@faouzi2022time;@li2020forecasting]. Inspired by `tsfresh`, we introduce `xr_fresh`, tailored specifically for gridded time series by automating the extraction of time series features on a pixel-by-pixel basis [@CHRIST201872].
 
 Currently, there is no method to rapidly extract a comprehensive set of features from gridded time series data, such as those derived from remote sensing imagery. Existing packages like `tsfresh` are not optimized for the unique structure of gridded time series data. This limitation hinders the ability to efficiently analyze and model these datasets, particularly in the context of remote sensing applications where large volumes of data are generated.
 
@@ -77,11 +77,11 @@ where $\vec{a}_{i,j} \in \mathbb{R}^U$ represents the $U$ univariate attributes 
 
 ### Time Series Feature Set
 
-The [online documentation](https://mmann1123.github.io/xr_fresh/feature_calculator_series.html) summarizes the suite of time series features extracted by the `xr_fresh` module from gridded data. These features are designed to characterize the temporal behavior of each pixel $(x_i, y_j)$. By including a diverse set of statistical, trend and distribution-based metrics, `xr_fresh` enables a detailed and scalable analysis of temporal patterns [@jin2022automated;@venkatachalam2024temporal]. Additional features can be added to the library as needed, and users can also define custom feature extraction functions.
+The [documentation](https://mmann1123.github.io/xr_fresh/feature_calculator_series.html) summarizes the suite of time series features extracted by the `xr_fresh` module from gridded data. These features are designed to characterize the temporal behavior of each pixel $(x_i, y_j)$. By including a diverse set of statistical, trend and distribution-based metrics, `xr_fresh` enables a detailed and scalable analysis of temporal patterns [@jin2022automated;@venkatachalam2024temporal]. Additional features can be added to the library as needed, and users can also define custom feature extraction functions.
 
-#### Interpolation
+### Interpolation
 
-The `xr_fresh` library includes functionality to interpolate missing values pixel-wise in the gridded data. The interpolation methods implemented in `xr_fresh` are designed to be computationally efficient and can handle large datasets effectively. The module supports advanced interpolation techniques including linear, nearest-neighbor, cubic, and univariate spline interpolation [@virtanen2020scipy]. These methods can utilize either regular intervals or provided date vectors. 
[…]
