### 2025-09-18 #9009 @editorialbot
[REVIEW]: xr_fresh: Automated Time Series Feature Extraction for Remote Sensing & Gridded Data
**Submitting author:** @mmann1123 (<a href="http://orcid.org/0000-0002-6268-6867">Michael Mann</a>)
**Repository:** https://github.com/mmann1123/xr_fresh
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.3
