### 2025-09-22 mmann1123/xr_fresh#35 @chenyangkang
[JOSS review] Simpler installation for geowombat
Thanks for this valuable package!

`xr_fresh` functionality depends on `geowombat` package, but I found the installation for `geowombat` somehow not smooth.

The `geowombat` package was not found on `micromamba` for some reason, and the installation got stagnant after I added the conda-forge channel. I eventually managed to install it using `pip install "geowombat@git+https://github.com/jgrss/geowombat.git`.

Also, would it be possible to push it to conda-forge? Although I'm not familiar with how C++ can be compatible there.

I can't speak for `conda` or `mamba` users since I did not install them. But it was worth labeling somewhere that if you have trouble installing geowombat you may do this.

Besides that, the installation for `xr_fresh` is smooth. Thanks.
