### 2025-09-22 mmann1123/xr_fresh#36 @chenyangkang
[JOSS review] Performance comparison
In the paper, the authors emphasized "rapidly extraction" of the features, which is a performance claim. Therefore it would be nice to have a performance comparison to extract those features using this package vs. using traditional methods in this field. Maybe one or two functions as samples for comparison will be enough.
