### 2025-09-22 mmann1123/xr_fresh#37 @chenyangkang
[JOSS review] Missing community guidelines
The author provided a valuable package but the community guidelines are missing for a third parties to

  1. Contribute to the software
  2. Report issues or problems with the software
  3. Seek support
