### 2025-09-22 #9009 @chenyangkang
Hi @mmann1123, thank you for putting together this valuable package. I have checked most of the items and left with a few minor questions.

1. I encountered an installation issue with`geowombat `, and that makes me wonder if adding additional documents or push it to conda-forge would be helpful (if possible). See https://github.com/mmann1123/xr_fresh/issues/35#issue-3442698906.
2. As a user I would be especially interested to see the comparison of `xr_fresh` with conventional spatial analysis pipeline, especially compared to using plain xarray + numpy. Speed seems to be a large advantage of the package so I'm missing this part. See https://github.com/mmann1123/xr_fresh/issues/36#issue-3442726269.
3. Community guidelines appear to be missing; please consider adding them. See https://github.com/mmann1123/xr_fresh/issues/37#issue-3442799846. For reference, adding [code of conduct](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project).
4. The references are overall well formatted and appropriate. I'm not sure what this one is referring to and why it is related: `Venkatachalam, S., Kacha, D., Sheth, D., Mann, M., & Jafari, A. (2024). Temporal patterns and pixel precision: Satellite-based crop classification using deep learning and machine learning. George Washington University, Department of Geography & Environment; Data Science Program.`
5. I also encountered a question about input data structure during the exploration: https://github.com/mmann1123/xr_fresh/issues/38#issue-3442861333

Overall, these issues should be considered minor and wouldn't merit withholding publication. I think performance comparison will be critical for a package emphasizing on computation speed, so I look forward to see the results.
