### 2025-09-24 mmann1123/xr_fresh#35 @chenyangkang
I just tested it gain and it works now using `micromamba install geowombat -c conda-forge`. The installation was stuck previously for some reason. Thanks.
