### 2025-09-24 fdc0b0c
#36 Add performance comparison
