### 2025-09-29 #9009 @mbobra
👋 @nvnsudharsan Thank you so much for agreeing to review! You can find the article in [this comment box](https://github.com/openjournals/joss-reviews/issues/9009#issuecomment-3309255470) from above ⬆️  and the software repository linked in the first comment box on this issue. To generate your checklist, use the following command:

I think you're good to go. Again, JOSS is an open review process and we encourage communication between the reviewers, the submitting author, and the editor. And please feel free to ask me questions, I'm always around.

Can you please respond here (or give a thumbs up) so I know you're in the right place and found all the materials?
