### 2025-10-07 5d5a3ba
resolves [JOSS review] Performance comparison Fixes #36
