### 2025-10-07 f922a55
[JOSS review] Input data structure  #38
