### 2025-10-07 3a5eb31
Netcdf (#39)
* init

* resolves [JOSS review] Performance comparison
Fixes #36

* [JOSS review] Input data structure
 #38
