### 2025-10-07 3bc1df6
[JOSS review] Missing community guidelines Fixes #37
