### 2025-10-07 5798022
[JOSS review] Missing community guidelines Fixes #37
