### 2025-10-07 a12a332
Revise README for GPU and NetCDF file support
Updated the README to improve clarity on GPU usage and added a section on working with NetCDF files.
