### 2025-10-20 fc71ac4
[JOSS review] Missing citation Fixes #41
