+++ writeup/refs.bib
@@ -69,17 +69,24 @@ abstract = {Time series feature engineering is a time-consuming process because 
   year={2022},
   doi={10.1109/TITS.2022.3195232},
   publisher={IEEE}
-}
-@article{lam2015numba,
-  title={Numba: A LLVM-based Python JIT Compiler},
-  author={Siu Kwan Lam and Antoine Pitrou and Stanley Seibert},
-  journal={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
-  year={2015},
-  pages={1--6},
-  doi={https://doi.org/10.1145/2833157.2833162},
-  url={https://numba.pydata.org/}
-}
- 
+} 
+@inproceedings{lam2015numba,
+  author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+  title = {Numba: a LLVM-based Python JIT compiler},
+  year = {2015},
+  isbn = {9781450340052},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/2833157.2833162},
+  doi = {10.1145/2833157.2833162},
+  abstract = {Dynamic, interpreted languages, like Python, are attractive for domain-experts and scientists experimenting with new ideas. However, the performance of the interpreter is often a barrier when scaling to larger data sets. This paper presents a just-in-time compiler for Python that focuses in scientific and array-oriented computing. Starting with the simple syntax of Python, Numba compiles a subset of the language into efficient machine code that is comparable in performance to a traditional compiled language. In addition, we share our experience in building a JIT compiler using LLVM[1].},
+  booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+  articleno = {7},
+  numpages = {6},
+  keywords = {LLVM, Python, compiler},
+  location = {Austin, Texas},
+  series = {LLVM '15}
+  }
 @article{li2020forecasting,
   title={Forecasting with time series imaging},
   author={Li, Xixi and Kang, Yanfei and Li, Feng},
