### 2025-10-20 #9009 @mmann1123
@mbobra Great, thanks for all your help. I think this has what you need https://zenodo.org/records/17401251
