### 2025-10-23 eaf3bde
Remove 'https://' from DOIs in refs.bib
