+++ writeup/refs.bib
@@ -7,7 +7,7 @@ volume = {307},
 pages = {72-77},
 year = {2018},
 issn = {0925-2312},
-doi = {https://doi.org/10.1016/j.neucom.2018.03.067},
+doi = {10.1016/j.neucom.2018.03.067},
 url = {https://www.sciencedirect.com/science/article/pii/S0925231218304843},
 author = {Maximilian Christ and Nils Braun and Julius Neuffer and Andreas W. Kempa-Liehr},
 keywords = {Feature engineering, Time series, Feature extraction, Feature selection, Machine learning},
@@ -105,7 +105,7 @@ volume = {201},
 pages = {60-74},
 year = {2017},
 issn = {0378-4290},
-doi = {https://doi.org/10.1016/j.fcr.2016.10.014},
+doi = {10.1016/j.fcr.2016.10.014},
 url = {https://www.sciencedirect.com/science/article/pii/S0378429016305238},
 author = {Michael L. Mann and James M. Warner},
 keywords = {Ethiopia, Agriculture, Data integration, Wheat productivity, Remote sensing, Smallholder agriculture, Panel data estimation, Yield gaps},
@@ -120,7 +120,7 @@ abstract = {Despite the routine collection of annual agricultural surveys and si
   number={1},
   pages={211--227},
   year={2019},
-  doi={10.1007/s10584-01902432-7},
+  doi={10.1007/s10584-019-02432-7},
   publisher={Springer}
 } 
 @article{mumuni2024automated,
