### 2025-10-23 d62bc1f
Update DOI link format in refs.bib
