+++ writeup/refs.bib
@@ -140,7 +140,7 @@ abstract = {Despite the routine collection of annual agricultural surveys and si
   journal={Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
-  doi={10.5555/1953048.2078195},
+  doi={https://dl.acm.org/doi/10.5555/1953048.2078195},
   year={2011}
 }
  
