### 2025-10-23 360527c
Fix DOI link format in refs.bib
