### 2025-10-23 414f213
Change DOI to URL format in refs.bib
Updated DOI to a new URL format for consistency.
