### 2025-11-01 #9009 @crvernon
@mmann1123 - Please correct the following before I accept this submission for publication:

In your paper:

- [ ] LINES 26-27:  Your link for "160 times longer" is not valid because you are linking to a path that will not exist to users who read this paper from JOSS.  See the raw declaration in your paper.md file:  `[160 times longer](../notebooks/time_trial.ipynb)`.  Please modify or remove this link.
- [ ] LINE 104:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing any characters you wish to maintain the formatting of in curly brackets.
- [ ] LINE 141:  The "e" in "ethiopia" should be capitalized.
- [ ] LINE 164:  The "p" in "python" should be capitalized.

Let me know when these changes have been made and after verifying them I'll accept.  Thanks!
