+++ xr_fresh/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 __author__ = "Michael Mann"
 __license__ = "MIT"
 __maintainer__ = "Michael Mann"
