+++ writeup/refs.bib
@@ -1,7 +1,7 @@
  
 
 @article{CHRIST201872,
-title = {Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh – A Python package)},
+title = {Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh A Python package)},
 journal = {Neurocomputing},
 volume = {307},
 pages = {72-77},
