### 2025-11-03 #9009 @mmann1123
@crvernon I updated the paper.md but those changes don't seem to be reflected when I ask generate pdf.  Maybe I am missing a step here.
