### 2025-11-03 60f2620
avoid numpy>2
