+++ .github/workflows/python-tests.yml
@@ -2,11 +2,10 @@ name: Python Unit Tests
 
 on:
   push:
-    branches: [ master ]
+    branches: [master]
   pull_request:
-    branches: [ master ]
-  workflow_dispatch:  # This line adds the manual trigger option
-
+    branches: [master]
+  workflow_dispatch: # This line adds the manual trigger option
 
 jobs:
   build:
@@ -14,63 +13,60 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [ '3.9', '3.10', '3.11']
+        python-version: ["3.9", "3.10", "3.11"]
 
     steps:
-    - name: Checkout code
-      uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
+      - name: Checkout code
+        uses: actions/checkout@v3
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
 
-    - name: Cache pip
-      uses: actions/cache@v3
-      with:
-        path: ~/.cache/pip
-        key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
-        restore-keys: |
-          ${{ runner.os }}-pip-
+      - name: Cache pip
+        uses: actions/cache@v3
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
 
+      - name: Install GDAL binaries
+        run: |
+          sudo apt update && sudo apt install -y software-properties-common
+          sudo add-apt-repository ppa:ubuntugis/ppa -y
+          sudo apt update && sudo apt install -y gdal-bin libgdal-dev libgl1 libspatialindex-dev g++ libmysqlclient-dev
+          echo "CPLUS_INCLUDE_PATH=/usr/include/gdal" >> $GITHUB_ENV
+          echo "C_INCLUDE_PATH=/usr/include/gdal" >> $GITHUB_ENV
+          echo "LD_LIBRARY_PATH=/usr/local/lib" >> $GITHUB_ENV
 
-    - name: Install GDAL binaries
-      run: |
-        sudo apt update && sudo apt install -y software-properties-common
-        sudo add-apt-repository ppa:ubuntugis/ppa -y
-        sudo apt update && sudo apt install -y gdal-bin libgdal-dev libgl1 libspatialindex-dev g++ libmysqlclient-dev
-        echo "CPLUS_INCLUDE_PATH=/usr/include/gdal" >> $GITHUB_ENV
-        echo "C_INCLUDE_PATH=/usr/include/gdal" >> $GITHUB_ENV
-        echo "LD_LIBRARY_PATH=/usr/local/lib" >> $GITHUB_ENV
+      - name: Verify GDAL Installation
+        run: which gdal-config
 
-    - name: Verify GDAL Installation
-      run: which gdal-config
-        
-    - name: Install Python packages
-      run: |
-        pip install -U pip setuptools wheel
-        pip install numpy
-        GDAL_VERSION=$(gdal-config --version | awk -F'[.]' '{print $1"."$2}')
-        pip install GDAL==$GDAL_VERSION --no-cache-dir
-        pip install -r requirements.txt
+      - name: Install Python packages
+        run: |
+          pip install -U pip setuptools wheel
+          GDAL_VERSION=$(gdal-config --version | awk -F'[.]' '{print $1"."$2}')
+          pip install GDAL==$GDAL_VERSION --no-cache-dir
+          pip install -r requirements.txt
 
-    - name: Install GeoWombat
-      run: |
[…]
