### 2025-11-03 ddf1729
force reinstall numpy version
