### 2025-11-03 dbb6248
update requirements
