+++ requirements.txt
@@ -4,7 +4,7 @@ h5netcdf>=0.8.0
 h5py>=2.10.0
 matplotlib>=3.1.3
 netcdf4>=1.5.3
-numpy>=1.18.0
+numpy>=1.18.0,<2.0.0
 pandas>=1.0.1
 rasterio>=1.1.2
 scipy>=1.4.1
