### 2025-11-03 4a464ee
move force reinstall
