+++ .github/workflows/python-tests.yml
@@ -48,7 +48,7 @@ jobs:
           pip install -U pip setuptools wheel
           pip install --force-reinstall 'numpy>=1.18.0,<2.0.0'
           GDAL_VERSION=$(gdal-config --version | awk -F'[.]' '{print $1"."$2}')
-          pip install GDAL==$GDAL_VERSION --no-cache-dir
+          pip install --force-reinstall GDAL==$GDAL_VERSION --no-cache-dir
           pip install -r requirements.txt
 
       - name: Install GeoWombat
