+++ writeup/refs.bib
@@ -1,7 +1,7 @@
  
 
 @article{CHRIST201872,
-title = {Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh A Python package)},
+title = {Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh A {P}ython package)},
 journal = {Neurocomputing},
 volume = {307},
 pages = {72-77},
@@ -113,7 +113,7 @@ abstract = {Despite the routine collection of annual agricultural surveys and si
 }
 
 @article{mann2019predicting,
-  title={Predicting high-magnitude, low-frequency crop losses using machine learning: an application to cereal crops in Ethiopia},
+  title={Predicting high-magnitude, low-frequency crop losses using machine learning: an application to cereal crops in {E}thiopia},
   author={Mann, Michael L and Warner, James M and Malik, Arun S},
   journal={Climatic change},
   volume={154},
@@ -163,7 +163,7 @@ abstract = {Despite the routine collection of annual agricultural surveys and si
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
