+++ LOCAL_TESTING.md
@@ -1,0 +1,254 @@
+# Local Testing Guide for xr_fresh
+
+This guide explains how to simulate the GitHub Actions CI/CD environment locally to test your code before pushing to GitHub.
+
+## Overview
+
+The GitHub Actions workflow (`.github/workflows/python-tests.yml`) tests your package across:
+- Python versions: 3.9, 3.10, 3.11
+- Ubuntu 22.04 environment
+- Specific versions of GDAL and other system dependencies
+
+You can simulate this locally using either **Docker** or **Conda environments**.
+
+---
+
+## Option 1: Testing with Docker (Recommended)
+
+Docker provides the most accurate simulation of the GitHub Actions environment.
+
+### Prerequisites
+- Docker installed on your system
+- At least 5GB of free disk space
+
+### Quick Test (Single Python Version)
+
+Test with a single Python version (default: 3.9):
+
+```bash
+# Build the Docker image
+docker build -f Dockerfile.test -t xr_fresh_test .
+
+# Run tests
+docker run --rm xr_fresh_test
+```
+
+### Test Specific Python Version
+
+```bash
+# Test with Python 3.10
+docker build -f Dockerfile.test -t xr_fresh_test:py3.10 .
+docker run --rm -e PYTHON_VERSION=3.10 xr_fresh_test:py3.10
+
+# Test with Python 3.11
+docker build -f Dockerfile.test -t xr_fresh_test:py3.11 .
+docker run --rm -e PYTHON_VERSION=3.11 xr_fresh_test:py3.11
+```
+
+### Test All Python Versions (Full Matrix)
+
+Run the complete test matrix across all Python versions:
+
+```bash
+chmod +x test-matrix.sh
+./test-matrix.sh
+```
+
+This will test Python 3.9, 3.10, and 3.11 sequentially, just like GitHub Actions.
+
+---
+
+## Option 2: Testing with Conda
+
+Conda is easier to set up if you're not familiar with Docker, but may not exactly match the GitHub Actions environment.
+
+### Prerequisites
+- Anaconda or Miniconda installed
+
+### Test All Python Versions
+
+```bash
+chmod +x test-with-conda.sh
+./test-with-conda.sh
+```
+
+This will:
+1. Create a fresh conda environment for each Python version
+2. Install all dependencies
+3. Run the unit tests
+4. Clean up the environment
+5. Report results
+
+### Manual Testing with Conda (Single Version)
+
+If you prefer manual control:
+
+```bash
+# Create environment
+conda create -n xr_fresh_test python=3.10 -y
+conda activate xr_fresh_test
+
+# Install dependencies
+conda install -c conda-forge gdal numpy -y
+pip install -r requirements.txt
+pip install "geowombat[perf,tests]@git+https://github.com/jgrss/geowombat.git"
+pip install testfixtures
+
+# Install package
+pip install -e .
+
+# Run tests
+python -m unittest discover -s tests -p 'test_*.py' -v
+
+# Cleanup
+conda deactivate
+conda env remove -n xr_fresh_test -y
+```
+
+---
+
+## Option 3: Testing in Your Current Environment
+
+If you just want to quickly run tests without isolation:
+
+```bash
+# Make sure you have the right dependencies
+pip install -r requirements.txt
+pip install "geowombat[perf,tests]@git+https://github.com/jgrss/geowombat.git"
+pip install testfixtures
+
+# Install package in editable mode
+pip install -e .
+
+# Run tests
+python -m unittest discover -s tests -p 'test_*.py' -v
+```
+
+**Note:** This approach might not catch environment-specific issues that would occur in GitHub Actions.
+
+---
+
+## Troubleshooting
+
+### NumPy Version Conflicts
+
[…]
