+++ xr_fresh/CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Mann
+  given-names: Michael L.
+  orcid: "https://orcid.org/0000-0002-6268-6867"
+doi: 10.5281/zenodo.17401251
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Mann
+    given-names: Michael L.
+    orcid: "https://orcid.org/0000-0002-6268-6867"
+  date-published: 2025-11-03
+  doi: 10.21105/joss.09009
+  issn: 2475-9066
+  issue: 115
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9009
+  title: "xr_fresh: Automated Time Series Feature Extraction for Remote
+    Sensing & Gridded Data"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09009"
+  volume: 10
+title: "xr_fresh: Automated Time Series Feature Extraction for Remote
+  Sensing & Gridded Data"
